S-HARDEN-1: pinned canonical content for SIL1, defence-in-depth int64 interpolation (F20/F22 partial)
Curie/Brunel paired, red-first: three new cells confirmed red against the
unmodified parser (got Ok, want BadSigmoidLutContent) before this fix landed —
a single off-by-one node, and the exact adjacent INT32_MIN/INT32_MAX operand
the external review's strike drove through this parser once already.

ParseSigmoidLut now validates every SIL1 node against the pinned canonical
table (include/superslm/silu_lut_canonical.h) byte-for-byte, after every
structural check and before any node is exposed through `out`. SIL1 is a
universal construction the spec fixes entirely, not model-specific learned
data, so content-pinning is the correct oracle (§17.3 cell 6) rather than
per-node range/monotonicity predicates, which a prior fix used and which the
strike showed do not close the class (F20's own record: "closes the entire
class" was withdrawn — this fix is stated as closing THIS operand, the SIL1
table content, not the KVC1/WSC1/ROP1 operands F22/F23/F24 name, which remain
unvalidated — see the parser-vs-consumer gate below).

SiluSigmoidQ15's interpolation intermediates (lo/hi/diff/product/delta) are
widened to int64_t as defence-in-depth beneath the content-pinning guarantee —
not because the canonical table needs it (the design's own int32 bound holds
under canonical content), but because F22's UB chain is the standing argument
against a downstream arithmetic step being the last line of defense on its own.

Also added: the F1 gate's literal red cell (a config-only v2 artifact expecting
a MissingSection diagnostic naming SigmoidLut specifically) and an end-to-end
cell running a hostile-content SIL1 through the REAL artifact/parser path
(OpenFromMemory, then ParseSigmoidLut on the section it returns) rather than
only the direct-SslmSectionView path the existing SIL1 suite used.

Build: MSVC build.bat, 7802 checks, 0 failures.

+++ src/silu_lut.cpp
@@ -49,15 +49,21 @@ int32_t SiluSigmoidQ15(const int32_t* table, int8_t code, int64_t m, int e) {
 	const int i0 = static_cast<int>(i0_64);
 	const int64_t frac = pos_fixed - (static_cast<int64_t>(i0) << kSiluLutQIdx);
 
-	// §6 — linear interpolation, exactly one C3 rounding (the shipped int32 primitive: the product
-	// fits int32, |frac·diff| <= 2^12 · 2^15 = 2^27). table[i0+1] always addresses a real entry
-	// (i0 <= N-1, and the table has N+1 entries for exactly this reason).
-	const int32_t lo = table[i0];
-	const int32_t hi = table[i0 + 1];
-	const int32_t diff = hi - lo;
-	const int32_t product = static_cast<int32_t>(frac * diff);
-	const int32_t delta = RoundingDivideByPOT(product, kSiluLutQIdx);
-	return lo + delta;
+	// §6 — linear interpolation, exactly one C3 rounding. table[i0+1] always addresses a real
+	// entry (i0 <= N-1, and the table has N+1 entries for exactly this reason). Under the
+	// canonical table (ParseSigmoidLut's pinned-content check, S-HARDEN-1 F20/F22 — the ONLY
+	// table this function is ever called with in the shipped path) |frac*diff| <= 2^12 * 2^15 =
+	// 2^27, which fits int32 exactly as the design derives. `diff`/`product`/`delta` are widened
+	// to int64_t as DEFENCE IN DEPTH beneath that content-pinning guarantee, not because the
+	// canonical table needs it: the guarantee is load-bearing at the parser, not re-derived here,
+	// but the two unvalidated-artifact-operand UB chain (F22) is the standing argument for never
+	// letting a downstream arithmetic step be the last line of defense on its own.
+	const int64_t lo = table[i0];
+	const int64_t hi = table[i0 + 1];
+	const int64_t diff = hi - lo;
+	const int64_t product = frac * diff;
+	const int64_t delta = RoundingDivideByPOT(product, kSiluLutQIdx);
+	return static_cast<int32_t>(lo + delta);
 }
 
 }  // namespace superslm
